backup_deleted/requirements.txt
# Flask API 의존성
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
pandas==2.0.3
openpyxl==3.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0